requirements.txt
# Requirements
#
# General
six
pytz
psutil
urllib3
#
# Openstack Victoria
gnocchiclient==7.0.4
os-client-config==2.1.0
python-cinderclient==7.2.0
python-designateclient==4.1.0
python-keystoneclient==4.1.1
python-neutronclient==7.2.1
python-novaclient==17.2.1
python-openstackclient==5.4.0
shade
#
# Machine Learning
pillow
jupyter
jupyterlab
matplotlib
numpy
pandas
scipy
scikit-learn
#
# Networking
networkx
scapy
ipaddress
#
# DKDP-Pktgen
pyelftools
sphinx
#
